+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 
 ## [Unreleased]
 
+## [2.10.0] - 2022-05-13
+
 ### Added
 
 - Allow to reuse any current context in `login` command by omitting the argument. This allows creating clientCerts for WCs in an arbitrary MC context. (not following `gs-codename` format)
@@ -869,7 +871,8 @@ This release supports rendering for CRs:
 - `AppCatalog`
 - `App`
 
-[Unreleased]: https://github.com/giantswarm/kubectl-gs/compare/v2.9.1...HEAD
+[Unreleased]: https://github.com/giantswarm/kubectl-gs/compare/v2.10.0...HEAD
+[2.10.0]: https://github.com/giantswarm/kubectl-gs/compare/v2.9.1...v2.10.0
 [2.9.1]: https://github.com/giantswarm/kubectl-gs/compare/v2.9.0...v2.9.1
 [2.9.0]: https://github.com/giantswarm/kubectl-gs/compare/v2.8.1...v2.9.0
 [2.8.1]: https://github.com/giantswarm/kubectl-gs/compare/v2.8.0...v2.8.1